+++ bundle/src/main/java/com/adobe/acs/commons/http/headers/impl/WeeklyExpiresHeaderFilter.java
@@ -19,59 +19,62 @@
  */
 package com.adobe.acs.commons.http.headers.impl;
 
-import org.apache.felix.scr.annotations.Component;
-import org.apache.felix.scr.annotations.ConfigurationPolicy;
-import org.apache.felix.scr.annotations.Properties;
-import org.apache.felix.scr.annotations.Property;
-import org.apache.felix.scr.annotations.PropertyOption;
-import org.apache.felix.scr.annotations.Service;
-import org.apache.sling.commons.osgi.PropertiesUtil;
-import org.osgi.service.cm.ConfigurationException;
-import org.osgi.service.component.ComponentContext;
 
 import java.util.Calendar;
 import java.util.Dictionary;
 
+import javax.servlet.Filter;
+
+import org.apache.sling.commons.osgi.PropertiesUtil;
+import org.osgi.service.cm.ConfigurationException;
+import org.osgi.service.component.ComponentContext;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.ConfigurationPolicy;
+import org.osgi.service.metatype.annotations.AttributeDefinition;
+import org.osgi.service.metatype.annotations.Designate;
+import org.osgi.service.metatype.annotations.ObjectClassDefinition;
+import org.osgi.service.metatype.annotations.Option;
+
 //@formatter:off
-@Component(
-    label = "ACS AEM Commons - Dispatcher Expires Header - Weekly",
-    description = "Adds an Expires header to content to enable Dispatcher TTL support.",
-    metatype = true,
-    configurationFactory = true,
-    policy = ConfigurationPolicy.REQUIRE)
-@Properties({
-  @Property(label = "Filter Patterns",
-      description = "Patterns on which to apply this Expires rule.",
-      cardinality = Integer.MAX_VALUE,
-      name = AbstractDispatcherCacheHeaderFilter.PROP_FILTER_PATTERN,
-      propertyPrivate = false,
-      value = { }),
-  @Property(label = "Expires Time",
-      description = "Time of day at which resources will expire. Must match SimpleDateFormat of 'HH:mm'.",
-      name = AbstractExpiresHeaderFilter.PROP_EXPIRES_TIME,
-      propertyPrivate = false),
-  @Property(
-        name = "webconsole.configurationFactory.nameHint",
-        value = "Expires Each week on day {expires.day-of-week} at {expires.time} for Patterns: [{filter.pattern}]",
-        propertyPrivate = true)
-})
+@Component(configurationPolicy=ConfigurationPolicy.REQUIRE,
+service=Filter.class,
+    factory = "WeeklyExpiresHeaderFilter", property= {
+    "webconsole.configurationFactory.nameHint" + "=" + "Expires Each week on day {expires.day-of-week} at {expires.time} for Patterns: [{filter.pattern}]"
+    })
+@Designate(ocd=WeeklyExpiresHeaderFilter.Config.class,factory=true)
 //@formatter:on
 public class WeeklyExpiresHeaderFilter extends AbstractExpiresHeaderFilter {
 
-    
-    @Property(
-            label = "Expires Day",
+@ObjectClassDefinition( name = "ACS AEM Commons - Dispatcher Expires Header - Weekly",
+    description = "Adds an Expires header to content to enable Dispatcher TTL support.")
+public @interface Config {
+
+  @AttributeDefinition(name = "Filter Patterns",
+      description = "Patterns on which to apply this Expires rule.",
+      cardinality = Integer.MAX_VALUE)
+  String[] filter_pattern();
+  
+  @AttributeDefinition(name = "Expires Time",
+      description = "Time of day at which resources will expire. Must match SimpleDateFormat of 'HH:mm'.")
+  String expires_time();
+
+  @AttributeDefinition(
+            name = "Expires Day",
             description = "Day of week on which content expires.",
             options = {
-                    @PropertyOption(name = "" + Calendar.SUNDAY, value = "Sunday"),
-                    @PropertyOption(name = "" + Calendar.MONDAY, value = "Monday"),
-                    @PropertyOption(name = "" + Calendar.TUESDAY, value = "Tuesday"),
-                    @PropertyOption(name = "" + Calendar.WEDNESDAY, value = "Wednesday"),
-                    @PropertyOption(name = "" + Calendar.THURSDAY, value = "Thursday"),
-                    @PropertyOption(name = "" + Calendar.FRIDAY, value = "Friday"),
-                    @PropertyOption(name = "" + Calendar.SATURDAY, value = "Saturday"),
+                    @Option(value = "" + Calendar.SUNDAY, label = "Sunday"),
+                    @Option(value = "" + Calendar.MONDAY, label = "Monday"),
+                    @Option(value = "" + Calendar.TUESDAY, label = "Tuesday"),
+                    @Option(value = "" + Calendar.WEDNESDAY, label = "Wednesday"),
+                    @Option(value = "" + Calendar.THURSDAY, label = "Thursday"),
+                    @Option(value = "" + Calendar.FRIDAY, label = "Friday"),
+                    @Option(value = "" + Calendar.SATURDAY, label = "Saturday"),
             })
-    static final String PROP_EXPIRES_DAY_OF_WEEK = "expires.day-of-week";
+    int expires_dayofweek();
+
+}
+
+static final String PROP_EXPIRES_DAY_OF_WEEK = "expires.dayofweek";
 
     private int dayOfWeek;
 
@@ -91,7 +94,7 @@ public class WeeklyExpiresHeaderFilter extends AbstractExpiresHeaderFilter {
         Dictionary<String, Object> props = context.getProperties();
         dayOfWeek = PropertiesUtil.toInteger(props.get(PROP_EXPIRES_DAY_OF_WEEK), -1);
         if (dayOfWeek < Calendar.SUNDAY || dayOfWeek > Calendar.SATURDAY) {
-            throw new ConfigurationException(PROP_EXPIRES_DAY_OF_WEEK, "Day of week must be valid value from Calendar DAY_OF_WEEK attribute.");
+            throw new ConfigurationException("expires.dayofweek", "Day of week must be valid value from Calendar DAY_OF_WEEK attribute.");
         }
     }
 }