+++ bundle/src/main/java/com/adobe/acs/commons/mcp/impl/processes/asset/FileAssetIngestor.java
@@ -171,7 +171,7 @@ public class FileAssetIngestor extends AssetIngestor {
         } catch (IOException ex) {
             Failure failure = new Failure();
             failure.setException(ex);
-            failure.setNodePath(fileSource.getElement().getNodePath());
+            failure.setNodePath(fileSource.getElement().getNodePath(preserveFileName));
             manager.getFailureList().add(failure);
         } finally {
             try {
@@ -179,7 +179,7 @@ public class FileAssetIngestor extends AssetIngestor {
             } catch (IOException ex) {
                 Failure failure = new Failure();
                 failure.setException(ex);
-                failure.setNodePath(fileSource.getElement().getNodePath());
+                failure.setNodePath(fileSource.getElement().getNodePath(preserveFileName));
                 manager.getFailureList().add(failure);
             }
         }