+++ bundle/src/main/java/com/adobe/acs/commons/packaging/impl/ACLPackagerServletImpl.java
@@ -20,6 +20,11 @@
 
 package com.adobe.acs.commons.packaging.impl;
 
+import static org.apache.sling.api.servlets.ServletResolverConstants.SLING_SERVLET_EXTENSIONS;
+import static org.apache.sling.api.servlets.ServletResolverConstants.SLING_SERVLET_METHODS;
+import static org.apache.sling.api.servlets.ServletResolverConstants.SLING_SERVLET_RESOURCE_TYPES;
+import static org.apache.sling.api.servlets.ServletResolverConstants.SLING_SERVLET_SELECTORS;
+
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -34,9 +39,8 @@ import java.util.regex.Pattern;
 
 import javax.jcr.RepositoryException;
 import javax.jcr.query.Query;
+import javax.servlet.Servlet;
 
-import org.apache.felix.scr.annotations.Reference;
-import org.apache.felix.scr.annotations.sling.SlingServlet;
 import org.apache.jackrabbit.api.security.user.Authorizable;
 import org.apache.jackrabbit.api.security.user.UserManager;
 import org.apache.jackrabbit.vault.fs.api.PathFilterSet;
@@ -48,6 +52,10 @@ import org.apache.sling.api.resource.ResourceResolver;
 import org.apache.sling.api.resource.ValueMap;
 import org.apache.sling.api.wrappers.ValueMapDecorator;
 
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+
 import com.adobe.acs.commons.packaging.PackageHelper;
 
 /**
@@ -55,12 +63,14 @@ import com.adobe.acs.commons.packaging.PackageHelper;
  * Servlet end-point used to create ACL CRX packages based on the underlying resource's configuration.
  */
 @SuppressWarnings("serial")
-@SlingServlet(
-        methods = { "POST" },
-        resourceTypes = { "acs-commons/components/utilities/packager/acl-packager" },
-        selectors = { "package" },
-        extensions = { "json" }
-)
+
+@Component(service=Servlet.class,
+property= {
+SLING_SERVLET_METHODS+"=POST",
+SLING_SERVLET_SELECTORS+"=package",
+SLING_SERVLET_EXTENSIONS+"=json",
+SLING_SERVLET_RESOURCE_TYPES+"=acs-commons/components/utilities/packager/acl-packager"
+})
 public class ACLPackagerServletImpl extends AbstractPackagerServlet {
 
     private static final String INCLUDE_PATTERNS = "includePatterns";