+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 - Add oakpal-maven-plugin and oakpal-checks module, using the acs-internal checklist for acs-aem-commons-content acceptance tests, and export the acs-commons-integrators checklist for downstream compatibility checks.
 - #1564 - Added SFTP support for asset ingest utilities
 - #1612 - Retries count and retry pause is configurable for all Asset Ingestors
+- #1637 - Add support for bounce address setting in EmailService
 
 ### Fixed
 - #1607 - HttpCache: improved the write to response mechanism.
@@ -32,6 +33,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 - #1610 - Bulk Workflow Manager doing nothing
 - #1613 - Potential NPE in JcrPackageReplicationStatusEventHandler
 - #1623 - Fix timing-related test failures in HealthCheckStatusEmailerTest
+- #1627 - Asset Ingestor and Valid Folder Name: if Preserve File name unchecked, asset and folder names will support only the following characters: letters, digits, hyphens, underscores, another chars will be replaced with hyphens
 
 ### Changed
 - #1571 - Remove separate twitter bundle and use exception trapping to only register AdapterFactory when Twitter4J is available.
@@ -39,7 +41,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 - #1578 - Asset import needs additional configuration inputs
 - #1615 - Add cq:Tag as a contentType for ContentVisitor API (allowing Content Traversing workflows to act upon cq:Tags)
 - #1609 - EnsureOakIndex excludes property seed, and sub-tree [oak:QueryIndexDefinition]/facets/jcr:content, by way up an updated to ChecksumGeneratorImpl that allows specific excludedNodeNames and excludedSubTrees.
-
+- #1560 - Changing to osgi annotations
+- #1634 - Made reference policy option greedy to allow plugging in a custom DispatcherFlusher service
 
 ## [3.19.0] - 2018-11-03